inotebook_backend/requirements.txt
Package                       Version
----------------------------- -------
asgiref                       3.6.0
Django                        4.1.7
django-cors-headers           3.14.0
django-dotenv                 1.4.2
djangorestframework           3.14.0
djangorestframework-simplejwt 5.2.2
pip                           20.2.3
PyJWT                         2.6.0
pytz                          2023.3
setuptools                    49.2.1
sqlparse                      0.4.3
tzdata                        2023.3
